refactor(library): implement asynchronous book addition and cover generation with event handling
