+++ booklore-api/src/main/java/org/booklore/BookloreApplication.java
@@ -3,13 +3,11 @@ package org.booklore;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 import org.booklore.config.BookmarkProperties;
 
 @EnableScheduling
-@EnableAsync
 @EnableConfigurationProperties(BookmarkProperties.class)
 @SpringBootApplication
 public class BookloreApplication {
