+++ booklore-api/src/main/java/org/booklore/service/watcher/LibraryFileEventProcessor.java
@@ -12,6 +12,7 @@ import org.booklore.repository.BookRepository;
 import org.booklore.repository.LibraryRepository;
 import org.booklore.service.file.FileFingerprint;
 import org.booklore.service.library.LibraryProcessingService;
+import org.booklore.service.library.LibraryScanListener;
 import org.booklore.util.FileUtils;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
@@ -45,6 +46,7 @@ public class LibraryFileEventProcessor {
     private final BookFileTransactionalHandler bookFileTransactionalHandler;
     private final BookFilePersistenceService bookFilePersistenceService;
     private final LibraryProcessingService libraryProcessingService;
+    private final LibraryScanListener libraryScanListener;
     private final PendingDeletionPool pendingDeletionPool;
 
     private final ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1);
@@ -70,6 +72,11 @@ public class LibraryFileEventProcessor {
     }
 
     public void processEvent(WatchEvent.Kind<?> eventKind, long libraryId, Path fullPath, boolean isDirectory) {
+        if (libraryScanListener.isScanning(libraryId)) {
+            log.debug("[SKIP] Library {} is currently being scanned, ignoring file event for '{}'", libraryId, fullPath);
+            return;
+        }
+
         Path path = fullPath.toAbsolutePath().normalize();
 
         if (eventKind == StandardWatchEventKinds.ENTRY_DELETE) {
