+++ booklore-api/src/test/java/org/booklore/service/library/FileAsBookProcessorTest.java
@@ -1,118 +1,49 @@
 package org.booklore.service.library;
 
-import org.booklore.model.FileProcessResult;
-import org.booklore.model.dto.Book;
-import org.booklore.model.dto.BookFile;
 import org.booklore.model.dto.settings.LibraryFile;
 import org.booklore.model.entity.LibraryEntity;
 import org.booklore.model.entity.LibraryPathEntity;
 import org.booklore.model.enums.BookFileType;
-import org.booklore.model.enums.FileProcessStatus;
-import org.booklore.repository.BookAdditionalFileRepository;
-import org.booklore.repository.BookRepository;
-import org.booklore.repository.LibraryRepository;
-import org.booklore.service.event.BookEventBroadcaster;
-import org.booklore.service.file.FileFingerprint;
-import org.booklore.service.fileprocessor.BookFileProcessor;
-import org.booklore.service.fileprocessor.BookFileProcessorRegistry;
-import org.booklore.service.kobo.KoboAutoShelfService;
-import org.booklore.service.metadata.extractor.AudiobookMetadataExtractor;
-import org.booklore.service.metadata.extractor.MetadataExtractorFactory;
-import org.booklore.util.FileService;
-import org.booklore.util.FileUtils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
-import org.mockito.MockedStatic;
 import org.mockito.MockitoAnnotations;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 class FileAsBookProcessorTest {
 
     @Mock
-    private BookEventBroadcaster bookEventBroadcaster;
-
-    @Mock
-    private BookFileProcessorRegistry processorRegistry;
-
-    @Mock
-    private KoboAutoShelfService koboAutoShelfService;
-
-    @Mock
-    private BookRepository bookRepository;
-
-    @Mock
-    private BookAdditionalFileRepository bookAdditionalFileRepository;
-
-    @Mock
-    private LibraryRepository libraryRepository;
-
-    @Mock
-    private FileService fileService;
-
-    @Mock
-    private MetadataExtractorFactory metadataExtractorFactory;
-
-    @Mock
-    private AudiobookMetadataExtractor audiobookMetadataExtractor;
-
-    @Mock
-    private BookFileProcessor bookFileProcessor;
+    private BookGroupProcessor bookGroupProcessor;
 
     private FileAsBookProcessor fileAsBookProcessor;
 
     private AutoCloseable mocks;
-    private MockedStatic<FileFingerprint> fileFingerprintMock;
-    private MockedStatic<FileUtils> fileUtilsMock;
 
     @BeforeEach
     void setUp() {
         mocks = MockitoAnnotations.openMocks(this);
-        fileAsBookProcessor = new FileAsBookProcessor(
-                bookEventBroadcaster,
-                processorRegistry,
-                koboAutoShelfService,
-                bookRepository,
-                bookAdditionalFileRepository,
-                libraryRepository,
-                fileService,
-                metadataExtractorFactory,
-                audiobookMetadataExtractor
-        );
-        fileFingerprintMock = mockStatic(FileFingerprint.class);
-        fileFingerprintMock.when(() -> FileFingerprint.generateHash(any(Path.class))).thenReturn("testhash");
-        fileUtilsMock = mockStatic(FileUtils.class);
-        fileUtilsMock.when(() -> FileUtils.getFileSizeInKb(any(Path.class))).thenReturn(100L);
-
+        fileAsBookProcessor = new FileAsBookProcessor(bookGroupProcessor);
     }
 
     @AfterEach
     void tearDown() throws Exception {
-        if (fileFingerprintMock != null) fileFingerprintMock.close();
-        if (fileUtilsMock != null) fileUtilsMock.close();
         if (mocks != null) mocks.close();
-    }
-
-    @Test
-    void staticMocksShouldWork() {
-        // Verify static mocks are working
-        Path testPath = Path.of("/test/path");
-        String hash = FileFingerprint.generateHash(testPath);
-        Long size = FileUtils.getFileSizeInKb(testPath);
-        assert hash.equals("testhash") : "Expected 'testhash' but got: " + hash;
-        assert size.equals(100L) : "Expected 100L but got: " + size;
     }
 
     @Test
     void processLibraryFiles_shouldProcessDifferentNamedFilesSeparately() {
         LibraryEntity libraryEntity = new LibraryEntity();
+        libraryEntity.setId(1L);
         LibraryPathEntity libraryPathEntity = new LibraryPathEntity();
         libraryPathEntity.setId(1L);
         libraryPathEntity.setPath("/library/path");
@@ -133,77 +64,26 @@ class FileAsBookProcessorTest {
                 .bookFileType(BookFileType.PDF)
                 .build();
 
-        Book book1 = Book.builder().id(1L).primaryFile(BookFile.builder().fileName("book1.epub").bookType(BookFileType.EPUB).build()).build();
-        Book book2 = Book.builder().id(2L).primaryFile(BookFile.builder().fileName("book2.pdf").bookType(BookFileType.PDF).build()).build();
-
-        when(processorRegistry.getProcessorOrThrow(BookFileType.EPUB)).thenReturn(bookFileProcessor);
-        when(processorRegistry.getProcessorOrThrow(BookFileType.PDF)).thenReturn(bookFileProcessor);
-        when(bookFileProcessor.processFile(file1)).thenReturn(new FileProcessResult(book1, FileProcessStatus.NEW));
-        when(bookFileProcessor.processFile(file2)).thenReturn(new FileProcessResult(book2, FileProcessStatus.NEW));
-
         fileAsBookProcessor.processLibraryFiles(List.of(file1, file2), libraryEntity);
 
-        verify(bookEventBroadcaster, times(2)).broadcastBookAddEvent(any());
-        verify(bookFileProcessor, times(2)).processFile(any());
+        // Files with different base names are in separate groups
+        verify(bookGroupProcessor, times(2)).process(any(), anyLong());
     }
 
     @Test
     void processLibraryFiles_shouldHandleEmptyList() {
         LibraryEntity libraryEntity = new LibraryEntity();
+        libraryEntity.setId(1L);
 
         fileAsBookProcessor.processLibraryFiles(new ArrayList<>(), libraryEntity);
 
-        verify(bookEventBroadcaster, never()).broadcastBookAddEvent(any());
-        verify(processorRegistry, never()).getProcessorOrThrow(any());
-    }
-
-    @Test
-    void processLibraryFiles_shouldSkipFilesWithNullBookFileType() {
-        LibraryEntity libraryEntity = new LibraryEntity();
-        LibraryPathEntity libraryPathEntity = new LibraryPathEntity();
-        libraryPathEntity.setId(1L);
-        libraryPathEntity.setPath("/library/path");
-
-        LibraryFile invalidFile = LibraryFile.builder()
-                .libraryEntity(libraryEntity)
-                .libraryPathEntity(libraryPathEntity)
-                .fileName("document.txt")
-                .fileSubPath("docs")
-                .bookFileType(null)
-                .build();
-
-        fileAsBookProcessor.processLibraryFiles(List.of(invalidFile), libraryEntity);
-
-        verify(bookEventBroadcaster, never()).broadcastBookAddEvent(any());
-        verify(processorRegistry, never()).getProcessorOrThrow(any());
-    }
-
-    @Test
-    void processLibraryFiles_shouldNotBroadcastWhenProcessorReturnsNull() {
-        LibraryEntity libraryEntity = new LibraryEntity();
-        LibraryPathEntity libraryPathEntity = new LibraryPathEntity();
-        libraryPathEntity.setId(1L);
-        libraryPathEntity.setPath("/library/path");
-
-        LibraryFile file = LibraryFile.builder()
-                .libraryEntity(libraryEntity)
-                .libraryPathEntity(libraryPathEntity)
-                .fileName("book.epub")
-                .fileSubPath("books")
-                .bookFileType(BookFileType.EPUB)
-                .build();
-
-        when(processorRegistry.getProcessorOrThrow(BookFileType.EPUB)).thenReturn(bookFileProcessor);
-        when(bookFileProcessor.processFile(file)).thenReturn(null);
-
-        fileAsBookProcessor.processLibraryFiles(List.of(file), libraryEntity);
-
-        verify(bookEventBroadcaster, never()).broadcastBookAddEvent(any());
+        verify(bookGroupProcessor, never()).process(any(), anyLong());
     }
 
     @Test
     void processLibraryFiles_shouldNotGroupFilesInDifferentDirectories() {
         LibraryEntity libraryEntity = new LibraryEntity();
+        libraryEntity.setId(1L);
         LibraryPathEntity libraryPathEntity = new LibraryPathEntity();
         libraryPathEntity.setId(1L);
         libraryPathEntity.setPath("/library/path");
@@ -224,18 +104,44 @@ class FileAsBookProcessorTest {
                 .bookFileType(BookFileType.PDF)
                 .build();
 
-        Book book1 = Book.builder().id(1L).primaryFile(BookFile.builder().fileName("book.epub").bookType(BookFileType.EPUB).build()).build();
-        Book book2 = Book.builder().id(2L).primaryFile(BookFile.builder().fileName("book.pdf").bookType(BookFileType.PDF).build()).build();
-
-        when(processorRegistry.getProcessorOrThrow(BookFileType.EPUB)).thenReturn(bookFileProcessor);
-        when(processorRegistry.getProcessorOrThrow(BookFileType.PDF)).thenReturn(bookFileProcessor);
-        when(bookFileProcessor.processFile(file1)).thenReturn(new FileProcessResult(book1, FileProcessStatus.NEW));
-        when(bookFileProcessor.processFile(file2)).thenReturn(new FileProcessResult(book2, FileProcessStatus.NEW));
-
         fileAsBookProcessor.processLibraryFiles(List.of(file1, file2), libraryEntity);
 
-        // Both should be processed as separate books
-        verify(bookFileProcessor, times(2)).processFile(any());
-        verify(bookEventBroadcaster, times(2)).broadcastBookAddEvent(any());
+        // Files in different directories should be processed separately
+        verify(bookGroupProcessor, times(2)).process(any(), anyLong());
+    }
+
+    @Test
+    void processLibraryFilesGrouped_shouldCallProcessForEachGroup() {
+        LibraryEntity libraryEntity = new LibraryEntity();
+        libraryEntity.setId(1L);
+
+        LibraryPathEntity libraryPathEntity = new LibraryPathEntity();
+        libraryPathEntity.setId(1L);
+        libraryPathEntity.setPath("/library/path");
+
+        LibraryFile file1 = LibraryFile.builder()
+                .libraryEntity(libraryEntity)
+                .libraryPathEntity(libraryPathEntity)
+                .fileName("book1.epub")
+                .fileSubPath("books")
+                .bookFileType(BookFileType.EPUB)
+                .build();
+
+        LibraryFile file2 = LibraryFile.builder()
+                .libraryEntity(libraryEntity)
+                .libraryPathEntity(libraryPathEntity)
+                .fileName("book2.pdf")
+                .fileSubPath("books")
+                .bookFileType(BookFileType.PDF)
+                .build();
+
+        Map<String, List<LibraryFile>> groups = Map.of(
+                "group1", List.of(file1),
+                "group2", List.of(file2)
+        );
+
+        fileAsBookProcessor.processLibraryFilesGrouped(groups, libraryEntity);
+
+        verify(bookGroupProcessor, times(2)).process(any(), anyLong());
     }
 }
