+++ booklore-api/src/test/java/org/booklore/service/library/LibraryProcessingServiceTest.java
@@ -49,6 +49,8 @@ class LibraryProcessingServiceTest {
     @Mock
     private BookGroupingService bookGroupingService;
     @Mock
+    private BookCoverGenerator bookCoverGenerator;
+    @Mock
     private EntityManager entityManager;
 
     private LibraryProcessingService libraryProcessingService;
@@ -64,6 +66,7 @@ class LibraryProcessingServiceTest {
                 bookDeletionService,
                 libraryFileHelper,
                 bookGroupingService,
+                bookCoverGenerator,
                 entityManager
         );
     }
